% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allBuildEnrichTable.R
\name{allBuildEnrichTable}
\alias{allBuildEnrichTable}
\title{Iterate \code{buildEnrichTable} along the specified GO ontologies and GO
levels}
\usage{
allBuildEnrichTable(
  x,
  geneUniverse,
  orgPackg,
  check.table = TRUE,
  ontos = c("BP", "CC", "MF"),
  GOLevels = GOLevels,
  keyType = "ENTREZID",
  storeEnrichedIn = TRUE,
  trace = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class "list". Each of its elements must be a "character"
vector of gene identifiers (e.g., ENTREZ). Then all pairwise contingency
tables of joint enrichment are built between these gene lists, iterating the
process for all specified GO ontologies and GO levels.}

\item{geneUniverse}{character vector containing the universe of genes from
where gene lists have been extracted. This vector must be obtained from the
annotation package declared in \code{orgPackg}. For more details, refer to
vignette \href{../doc/goSorensen_Introduction.html}{goSorensen_Introduction}.}

\item{orgPackg}{A string with the name of the genomic annotation package
corresponding to a specific species to be analysed, which must be previously
installed and activated. For more details, refer to vignette
\href{../doc/goSorensen_Introduction.html}{goSorensen_Introduction}.}

\item{check.table}{Boolean. If TRUE (default), all resulting tables are
checked by means of function \code{nice2x2Table}.}

\item{ontos}{"character", GO ontologies to analyse. Defaults to \code{c("BP",
"CC", "MF")}.}

\item{GOLevels}{Integer vector of GO levels to analyze inside each selected
ontology.
  If \code{NULL}, contingency tables are built without restricting the
  analysis to specific GO levels.}

\item{keyType}{Character string specifying the type of gene identifier used
in the input, such as \code{"ENTREZID"} or \code{"SYMBOL"}.}

\item{storeEnrichedIn}{logical, for each ontology and level under study, the
matrix of enriched (GO terms) x (gene lists) TRUE/FALSE values, must be
stored in the result?}

\item{trace}{Logical. If TRUE (default), the (usually very time consuming)
process of function
\code{allbuildEnrichTable} is traced along the specified GO ontologies and
levels.}

\item{...}{extra parameters for function \code{buildEnrichTable}.}
}
\value{
An object of class "allTableList". It is a list with as many
components as GO ontologies have been analyzed.

  If \code{GOLevels} is not \code{NULL}, each ontology component is itself a
  list with as many components as GO levels have been analyzed. Each of these
  elements is an object generated by
  \code{buildEnrichTable.list()}, that is, an object of class "tableList"
  containing all pairwise contingency tables of mutual enrichment between the
  gene lists in argument \code{x}.

  If \code{GOLevels = NULL}, each ontology component is directly an object of
  class "tableList", containing all pairwise contingency tables of mutual
  enrichment between the gene lists in argument \code{x}, without GO level
  restriction.

  The result also includes the attribute \code{GOLevels}, which stores the GO
  levels used in the analysis, or \code{NULL} when no GO level restriction
  was applied.
}
\description{
Iterate \code{buildEnrichTable} along the specified GO ontologies and GO
levels
}
\examples{
## The following example is highly time-consuming and is therefore not run
## automatically during R CMD check.

\dontrun{
## i) Obtaining ENTREZ identifiers for the gene universe of humans:
library(org.Hs.eg.db)
humanEntrezIDs <- keys(org.Hs.eg.db, keytype = "ENTREZID")

## ii) Gene lists to be explored for analysis:
data(allOncoGeneLists)

# iii) Calculation of all joint enrichment contingency matrices directly from
# the gene lists across GO levels 3 to 10 for the BP, CC, and MF ontologies.
allContTabs <- allBuildEnrichTable(allOncoGeneLists,
  geneUniverse = humanEntrezIDs,
  orgPackg = "org.Hs.eg.db",
  ontos = c("BP", "CC", "MF"),
  GOLevels = 3:10
)
allContTabs
}

# Since running this example may take several minutes, the result has been
# pre-computed and is accessible as the following:
data(allContTabs)
allContTabs
# This shortcut applies only to this example; for your own gene-list data,
# the computation must be performed explicitly.

# For a complete overview of this function's use, see the section 3 of the
# vignette "Introduction to goSorensen". You can do this by consulting the
# general package documentation or by directly running the following code in
# the R console:
# vignette("goSorensen_Introduction", package = "goSorensen")

}
