% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allDissMatrx.R
\docType{data}
\name{allDissMatrx}
\alias{allDissMatrx}
\title{Example of the output produced by the function \code{allSorenThreshold}. It contains the dissimilarity matrices for GO levels from 3 to 10 across the ontologies BP, CC and MF.}
\format{
An object of class "dist"
}
\usage{
data("allDissMatrx")
}
\description{
This object contains the matrices of dissimilarities between the 7 lists from \code{allOncoGeneLists}, computed based on the irrelevance threshold that makes them equivalent for GO levels from 3 to 10 across the ontologies BP, CC and MF.
}
\details{
Equivalence tests were computed based on the normal distribution (\code{boot = TRUE} by default) and using a confidence level \code{conf.level = 0.95}.

Consider this object only as an illustrative example, which is valid exclusively for the data \code{\link{allOncoGeneLists}} contained in this package. Note that gene lists, GO terms, and Bioconductor may change over time. The current version of these results were generated with Bioconductor version 3.20.
}
\keyword{datasets}
