% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEqTestsNoLevel.R
\docType{data}
\name{allEqTestsNoLevel}
\alias{allEqTestsNoLevel}
\title{Example of the output produced by the function \code{allEquivTestSorensen} using the normal asymptotic distribution.}
\format{
An exclusive object from \code{goSorensen} of the class "AllEquivSDhtest"
}
\usage{
data(allEqTestsNoLevel)
}
\description{
This object contains all the outputs for the equivalence tests to compare all possible pairs of lists from \code{\link{allOncoGeneLists}} without GO-Level restriction, and for the ontologies BP, CC, and MF, using the normal asymptotic distribution.
}
\details{
The parameters considered to execute these tests are: irrelevance limit \code{d0 = 0.4444} and confidence level \code{conf.level = 0.95}.

Consider this object only as an illustrative example, which is valid exclusively for the data \code{\link{allOncoGeneLists}} contained in this package. Note that gene lists, GO terms, and Bioconductor may change over time. The current version of these results were generated with Bioconductor version 3.20.
}
\keyword{datasets}
