% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEquivTestSorensen.R
\name{allEquivTestSorensen}
\alias{allEquivTestSorensen}
\alias{allEquivTestSorensen.list}
\alias{allEquivTestSorensen.allTableList}
\title{Iterate \code{equivTestSorensen} along the specified GO ontologies and GO
levels}
\usage{
allEquivTestSorensen(x, ...)

\method{allEquivTestSorensen}{list}(
  x,
  d0 = 1/(1 + 1.25),
  conf.level = 0.95,
  boot = FALSE,
  nboot = 10000,
  check.table = TRUE,
  ontos = c("BP", "CC", "MF"),
  GOLevels = GOLevels,
  trace = TRUE,
  ...
)

\method{allEquivTestSorensen}{allTableList}(
  x,
  d0 = 1/(1 + 1.25),
  conf.level = 0.95,
  boot = FALSE,
  nboot = 10000,
  check.table = TRUE,
  ontos,
  GOLevels,
  trace = TRUE,
  ...
)
}
\arguments{
\item{x}{either an object of class "list" or an object of class
"allTableList". In the first case, each of its elements must be a "character"
vector of gene identifiers (e.g., ENTREZ).}

\item{...}{extra parameters for function \code{buildEnrichTable}.}

\item{d0}{equivalence threshold for the Sorensen-Dice dissimilarity, d.
The null hypothesis states that d >= d0, i.e., inequivalence between the
compared gene lists and the alternative that d < d0, i.e., equivalence or
dissimilarity irrelevance (up to a level d0).}

\item{conf.level}{confidence level of the one-sided confidence interval, a
value between 0 and 1.}

\item{boot}{boolean. If TRUE, the confidence interval and the test p-value
are computed by means of a bootstrap approach instead of the asymptotic
normal approach. Defaults to FALSE.}

\item{nboot}{numeric, number of initially planned bootstrap replicates.
Ignored if \code{boot == FALSE}. Defaults to 10000.}

\item{check.table}{Boolean. If TRUE (default), argument \code{x} is checked
to adequately represent a 2x2 contingency table (or an aggregate of them) or
gene lists producing a correct table. This checking is performed by means of
function \code{nice2x2Table}.}

\item{ontos}{"character", GO ontologies to analyse. Defaults to
\code{c("BP", "CC", "MF")}.}

\item{GOLevels}{Integer vector of GO levels to analyze inside each selected
ontology.
  If \code{NULL}, equivalence tests are performed without restricting the
  analysis to specific
  GO levels.}

\item{trace}{Logical. If TRUE (default), the (usually very time consuming)
process of function
\code{allEquivTestSorensen} is traced along the specified GO ontologies and
levels.}
}
\value{
An object of class "AllEquivSDhtest". It is a list with as many components as
GO ontologies have been analyzed.

If \code{GOLevels} is not \code{NULL}, each ontology component is itself a
list with as many components as GO levels have been analyzed. Each of these
elements is an object generated by
\code{equivTestSorensen.list()} or \code{equivTestSorensen.tableList()}, that
is, an object of class "equivSDhtestList" containing pairwise comparisons
between gene lists.

If \code{GOLevels = NULL}, each ontology component is directly an object of
class "equivSDhtestList", containing all pairwise equivalence tests between
the gene lists in argument \code{x}, without GO level restriction.
}
\description{
Iterate \code{equivTestSorensen} along the specified GO ontologies and GO
levels
}
\section{Methods (by class)}{
\itemize{
\item \code{allEquivTestSorensen(list)}: S3 method for class "list"

\item \code{allEquivTestSorensen(allTableList)}: S3 method for class "allTableList"

}}
\examples{
## The following example is highly time-consuming and is therefore not run
## automatically during R CMD check.

\dontrun{
## i) Obtaining ENTREZ identifiers for the gene universe of humans:
library(org.Hs.eg.db)
humanEntrezIDs <- keys(org.Hs.eg.db, keytype = "ENTREZID")

## ii) Gene lists to be explored for analysis:
data(allOncoGeneLists)

# iii) Calculation of equivalence tests for all joint enrichment contingency
# tables across GO levels 3 to 10 for the BP, CC, and MF ontologies
allEqTests <- allEquivTestSorensen(allOncoGeneLists,
  geneUniverse = humanEntrezIDs,
  orgPackg = "org.Hs.eg.db",
  ontos = c("BP", "CC", "MF"),
  GOLevels = 3:10,
  d0 = 0.4444,
  conf.level = 0.95
)
allEqTests
}

# Since running this example may take several minutes, the result has been
# pre-computed and is accessible as the following:
data(allEqTests)
allEqTests
# This shortcut applies only to this example; for your own gene-list data,
# the computation must be performed explicitly.

# For a complete overview of this function's use, see the section 4 of the
# vignette "Introduction to goSorensen". You can do this by consulting the
# general package documentation or by directly running the following code in
# the R console:
# vignette("goSorensen_Introduction", package = "goSorensen")

}
