% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allOncoGeneLists.R
\docType{data}
\name{allOncoGeneLists}
\alias{allOncoGeneLists}
\title{7 gene lists possibly related with cancer}
\format{
An object of class "list" of length 7. Each one of its elements is a "character" vector
of gene identifiers.
}
\source{
\url{http://www.bushmanlab.org/links/genelists}
}
\usage{
allOncoGeneLists
}
\description{
An object of class "list" of length 7. Each one of its elements is a "character" vector
of gene identifiers. Only gene lists of length almost 100 were taken from their source web.
Take these lists just as an illustrative example, they are not automatically updated.
}
\keyword{datasets}
