% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allTabsBP.4.R
\docType{data}
\name{allTabsBP.4}
\alias{allTabsBP.4}
\title{An example of an object of class "tableList" resulting from a call to 'buildEnrichTable'}
\format{
An object of class "tableList" inheriting from class "list". It is a list of class
"table" objects.
}
\usage{
data(allTabsBP.4)
}
\description{
The result of generating all contingency tables of mutual enrichment, in a pairwise fashion,
between the gene lists in data \code{\link{allOncoGeneLists}}. The information in these data
was summarized as 2x2 contingency tables of GO terms enrichment, at level 4 of the BP
ontology. These results are based on gene lists which are not automatically updated, take
them just as an illustrative example because the gene lists, the GO and Bioconductor, may change
along time. The present version of these data was generated under Bioconductor version 3.17.
}
\keyword{datasets}
