% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerEquivSorensen.R, R/pbtEquivSorensen.R
\docType{data}
\name{cancerEquivSorensen}
\alias{cancerEquivSorensen}
\title{The Sorensen-Dice test performed on the cancer gene lists in data \code{\link{allOncoGeneLists}} which
is automatically charged with this package. The test is iterated for all GO ontologies and for GO levels 3 to 10.}
\format{
An object of class "AllEquivSDhtest" inheriting from class "list". Each one of its elements, named
BP, CC and MF respectively, corresponds to a GO ontology. It is itself a list of length 8 whose elements
are named as "Level 3" to "Level 10". For each combination of ontology and level, there is an object of
class "equivSDhtestList" codifying the result of all pairwise tests between these cancer gene lists.

An object of class "AllEquivSDhtest" inheriting from class "list". Each one of its elements, named
BP, CC and MF respectively, corresponds to a GO ontology. It is itself a list of length 8 whose elements
are named as "Level 3" to "Level 10". For each combination of ontology and level, there is an object of
class "equivSDhtestList" codifying the result of all pairwise tests between these kidney rejection gene
lists.
}
\source{
\url{http://www.bushmanlab.org/links/genelists}
}
\usage{
cancerEquivSorensen

cancerEquivSorensen
}
\description{
For each ontology and GO level, the result contains the result of all pairwise tests of equivalence between
the cancer gene lists.

For each ontology and GO level, the result contains the result of all pairwise tests of equivalence between
these gene lists.
}
\examples{
# This code may help to understand the structure of these data:
names(pbtAllOntosAndLevels)
names(pbtAllOntosAndLevels$BP)
names(pbtAllOntosAndLevels$BP$`level 4`)
class(pbtAllOntosAndLevels$BP$`level 4`)
pbtAllOntosAndLevels$BP$`level 4`
names(pbtAllOntosAndLevels$BP$`level 4`$KT1)
class(pbtAllOntosAndLevels$BP$`level 4`$KT1)
class(pbtAllOntosAndLevels$BP$`level 4`$KT1$IRITD5)
pbtAllOntosAndLevels$BP$`level 4`$KT1$IRITD5
}
\keyword{datasets}
