% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeTable.R
\name{completeTable}
\alias{completeTable}
\title{Reformats and completes (if necessary) an enrichment contingency table as is generated
by function 'crossTabGOIDs4GeneLists', in order to make it appropriate for its use in
package goSorensen.}
\usage{
completeTable(x, listNames)
}
\arguments{
\item{x}{an object of class "table", typically the output of function 'crossTabGOIDs4GeneLists'.}

\item{listNames}{a character(2) with the gene lists names originating the cross-tabulated
enrichment frequencies.}
}
\value{
a complete contingency table to use in package goSorensen.
}
\description{
Reformats and completes (if necessary) an enrichment contingency table as is generated
by function 'crossTabGOIDs4GeneLists', in order to make it appropriate for its use in
package goSorensen.
}
