% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_atlas.sanger_BP4.R
\docType{data}
\name{cont_atlas.sanger_BP4}
\alias{cont_atlas.sanger_BP4}
\title{Example of the output produced by the function \code{buildEnrichTable}. It contains the enrichment contingency table for two lists at level 4 of ontology BP.}
\format{
An object of class "table"
}
\usage{
data(cont_atlas.sanger_BP4)
}
\description{
A contingency 2x2 table with the number of joint enriched GO terms (TRUE-TRUE); the number of GO terms enriched only in one list but not in the other one (FALSE-TRUE and TRUE-FALSE); and the number of GO terms not enriched in either of the two lists.
}
\details{
Consider this object only as an illustrative example, which is valid exclusively for the lists atlas and sanger from the data \code{\link{allOncoGeneLists}} contained in this package. Note that gene lists, GO terms, and Bioconductor may change over time. The current version of these results were generated with Bioconductor version 3.20.
}
\keyword{datasets}
