% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTest.R
\name{crossTabGOIDs}
\alias{crossTabGOIDs}
\title{crossTabGOIDs}
\usage{
crossTabGOIDs(
  GO1,
  GO2,
  onto,
  GOLev,
  listNames = NULL,
  geneList1 = NULL,
  geneList2 = NULL,
  orgPackage = NULL,
  restricted = FALSE
)
}
\arguments{
\item{GO1}{character vector containing a FIRST list of GO identifiers}

\item{GO2}{character vector containing a SECOND gene list GO identifiers}

\item{onto}{string describing the ontology. Belongs to c('BP', 'MF', 'CC', 'ANY')}

\item{GOLev}{An integer}

\item{listNames}{character vector with names of the genelists that generated the enriched GOIDs}

\item{geneList1}{character vector containing a FIRST gene list of entrez IDs}

\item{geneList2}{character vector containing a SECOND gene list of entrez IDs}

\item{orgPackage}{A string wih the name of the annotation package}

\item{restricted}{Boolean variable to decide how tabulation is performed.
Unrestricted tabulation crosses _all_ GO Terms located at the level indicated by `GOLev` with the two GOIDs lists
Restricted tabulation crosses only terms from the selected GO level that are _common to ancestor terms of either list_.
That is, if one term in the selected GO level is not an ancestor of at least one of the gene list most specific GO terms
it is excluded from the GO Level's terms because it is impossible that it appears as being enriched.}
}
\description{
This function performs a crosstabulation between two lists of enriched GOTerms
The lists are intended to have been obtained from enrichment analyses performed on two gene lists
}
\examples{
 data(kidneyGeneLists)
 data(kidneyEnrichedGOIDs)
 GOIDs1 <-kidneyEnrichedGOIDs[[1]]
 GOIDs2 <-kidneyEnrichedGOIDs[[2]]
 names4lists<- names(kidneyEnrichedGOIDs)[1:2]
 anOnto <- 'BP'
 GOLev <- 3
 gl1 <- kidneyGeneLists[[1]]
 gl2 <- kidneyGeneLists[[2]]
 orgPkg ='org.Hs.eg.db'
 restrictTab <- TRUE
 crossTabbedGOIdsRestricted <- crossTabGOIDs (GO1 = GOIDs1, GO2 = GOIDs2, onto = anOnto,
 GOLev =GOLev, listNames=names4lists, geneList1 = gl1, geneList2 = gl2, orgPackage =orgPkg, restricted = restrictTab)
 show(crossTabbedGOIdsRestricted)
 restrictTab <- FALSE
 crossTabbedGOIdsUnrestricted <- crossTabGOIDs (GO1 = GOIDs1, GO2 = GOIDs2, onto = anOnto,
 GOLev =GOLev, listNames=names4lists, geneList1 = gl1, geneList2 = gl2, orgPackage =orgPkg, restricted = restrictTab)
 show(crossTabbedGOIdsUnrestricted)
}
