% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTest.R
\name{crossTabGOIDs4GeneLists}
\alias{crossTabGOIDs4GeneLists}
\title{crossTab4GeneLists}
\usage{
crossTabGOIDs4GeneLists(
  genelist1,
  genelist2,
  geneUniverse,
  orgPackg,
  onto,
  GOLev,
  restricted = FALSE,
  pAdjustMeth = "BH",
  pvalCutoff = 0.01,
  qvalCutoff = 0.05
)
}
\arguments{
\item{genelist1}{character vector containing a FIRST gene list of entrez IDs}

\item{genelist2}{character vector containing a SECOND gene list of entrez IDs}

\item{geneUniverse}{character vector containing all genes from where geneLists have been extracted}

\item{orgPackg}{A string wih the name of the annotation package}

\item{onto}{string describing the ontology. Belongs to c('BP', 'MF', 'CC', 'ANY')}

\item{GOLev}{An integer}

\item{restricted}{Boolean variable to decide how tabulation of GOIDs is performed.
Unrestricted tabulation crosses _all_ GO Terms located at the level indicated by `GOLev` with the two GOIDs lists
Restricted tabulation crosses only terms from the selected GO level that are _common to ancestor terms of either list_.
That is, if one term in the selected GO level is not an ancestor of at least one of the gene list most specific GO terms
it is excluded from the GO Level's terms because it is impossible that it appears as being enriched.}

\item{pAdjustMeth}{string describing the adjust method. Belongs to c('BH', 'BY', 'Bonf')}

\item{pvalCutoff}{A numeric value}

\item{qvalCutoff}{A numeric value}
}
\description{
This function builds a cross-tabulation of enriched and non-enriched GO terms
from two gene lists
}
\examples{
 data(kidneyGeneLists)
 data(humanEntrezIDs)
 gl1 <-kidneyGeneLists[[1]]
 gl2 <-kidneyGeneLists[[2]]
 anOnto <- 'BP'
 GOLev<- 3
 restricted <- FALSE
 adjMeth<- 'BH'
 pValCut <- 0.05
 qValCut <- 0.01
 crossTabFromGeneListsUnrestricted <- crossTabGOIDs4GeneLists (genelist1=gl1, genelist2=gl2,
 geneUniverse=humanEntrezIDs, orgPackg="org.Hs.eg.db", onto=anOnto, GOLev=GOLev, restricted=restricted)
 restricted <- TRUE
 crossTabFromGeneListsRestricted <- crossTabGOIDs4GeneLists (genelist1=gl1, genelist2=gl2,
 geneUniverse=humanEntrezIDs, orgPackg="org.Hs.eg.db", onto=anOnto, GOLev=GOLev, restricted=restricted)
}
