% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTest.R
\name{crossTabGOIDsUnrestricted}
\alias{crossTabGOIDsUnrestricted}
\title{crossTabGOIDsUnrestricted}
\usage{
crossTabGOIDsUnrestricted(GO1, GO2, onto, GOLev, listNames = NULL)
}
\arguments{
\item{GO1}{character vector containing a FIRST list of GO identifiers}

\item{GO2}{character vector containing a SECOND gene list GO identifiers}

\item{onto}{string describing the ontology. Belongs to c('BP', 'MF', 'CC', 'ANY')}

\item{GOLev}{An integer}

\item{listNames}{character vector with names of the genelists that generated the
enriched GOIDs}
}
\value{
a crosstabulation between two lists of enriched GOTerms
}
\description{
This function performs a crosstabulation between two lists of enriched GOTerms
The lists are intended to have been obtained from enrichment analyses performed
on two gene lists
}
