% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTest.R
\name{enrichOnto}
\alias{enrichOnto}
\title{enrichOnto}
\usage{
enrichOnto(
  geneList,
  geneUniverse,
  orgPackage = "org.Hs.eg.db",
  onto = c("BP", "MF", "CC"),
  pAdjustMeth = "BH",
  pvalCutoff = 0.01,
  qvalCutoff = 0.05
)
}
\arguments{
\item{geneList}{character vector containing a FIRST gene list of entrez IDs}

\item{geneUniverse}{character vector containing all genes from where geneLists have been extracted}

\item{orgPackage}{A string wih the name of the annotation package}

\item{onto}{string describing the ontology. Belongs to c('BP', 'MF', 'CC', 'ANY')}

\item{pAdjustMeth}{string describing the adjust method. Belongs to c('BH', 'BY', 'Bonf')}

\item{pvalCutoff}{A numeric value}

\item{qvalCutoff}{A numeric value}
}
\description{
This function performs standard tests of enrichment from a gene list
}
