% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTest.R
\name{extractGOIDs}
\alias{extractGOIDs}
\title{extractGOIDs}
\usage{
extractGOIDs(enriched)
}
\arguments{
\item{enriched}{An object with the output of an enricgment analysis performed with clusterprofiler package}
}
\description{
This is an auxiliary function to extract GOIDs from the output of an enricgment analysis performed with clusterprofiler package
}
