% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDissimilarity.R
\name{getDissimilarity}
\alias{getDissimilarity}
\alias{getDissimilarity.equivSDhtest}
\alias{getDissimilarity.equivSDhtestList}
\alias{getDissimilarity.AllEquivSDhtest}
\title{Access to the estimated Sorensen-Dice dissimilarity in one or more e
quivalence test results}
\usage{
getDissimilarity(x, ...)

\method{getDissimilarity}{equivSDhtest}(x, ...)

\method{getDissimilarity}{equivSDhtestList}(x, simplify = TRUE, ...)

\method{getDissimilarity}{AllEquivSDhtest}(x, onto, GOLevel, listNames, simplify = TRUE, ...)
}
\arguments{
\item{x}{an object of class "equivSDhtest" or "equivSDhtestList" or
"allEquivSDtest".}

\item{...}{Additional parameters.}

\item{simplify}{logical, if TRUE the result is simplified, e.g., returning a
vector instead
of a matrix.}

\item{onto}{character, a vector with one or more of "BP", "CC" or "MF",
ontologies to access.}

\item{GOLevel}{numeric or character, a vector with one or more GO levels to
access.
See the details section and the examples.}

\item{listNames}{character(2), the names of a pair of gene lists.}
}
\value{
When \code{x} is an object of class "equivSDhtest" (i.e., the result
of a single equivalence test), the returned value is a single numeric value,
the Sorensen-Dice dissimilarity. For an object of class "equivSDhtestList"
(i.e. all pairwise tests for a set of gene lists), if
\code{simplify = TRUE} (the default), the resulting value is a vector
with the dissimilarities in all those tests, or the symmetric matrix of all
dissimilarities if \code{simplify = TRUE}. If \code{x} is an object of class
"allEquivSDtest" (i.e., the test iterated along GO ontologies and levels),
the preceding result is returned in the form of a list along the ontologies,
levels and pairs of gene lists specified by the arguments
\code{onto, GOlevel} and \code{listNames} (or all present in \code{x} for
missing arguments).
}
\description{
Given objects representing the result(s) of one or more equivalence tests
(classes "equivSDhtest", "equivSDhtestList" or "allEquivSDtest", i.e., the
result of functions 'equivTestSorensen' and 'allEquivTestSorensen')
this function returns the estimated dissimilarities in the tests.
}
\details{
Argument \code{GOLevel} can be of class "character" or "numeric". In the
first case, the GO levels must be specified like \code{"level 6"} or
\code{c("level 4", "level 5", "level 6")} In the second case ("numeric"),
the GO levels must be specified like\code{6} or \code{seq.int(4,6)}.
}
\section{Methods (by class)}{
\itemize{
\item \code{getDissimilarity(equivSDhtest)}: S3 method for class "equivSDhtest"

\item \code{getDissimilarity(equivSDhtestList)}: S3 method for class "equivSDhtestList"

\item \code{getDissimilarity(AllEquivSDhtest)}: S3 method for class "AllEquivSDhtest"

}}
\examples{
# Manually define a 2 x 2 enrichment contingency table
contTable <- as.table(matrix(c(127, 19, 159, 3018),
  nrow = 2,
  dimnames = list(
    "Enriched in List 1" = c(TRUE, FALSE),
    "Enriched in List 2" = c(TRUE, FALSE)
  )
))
contTable

# Compute the Sorensen equivalence test from the contingency table.
# The result is an object of class "equivSDhtest".
equivalenceTest <- equivTestSorensen(contTable)
equivalenceTest

# Extract the Sorensen dissimilarity value from the equivalence test object
getDissimilarity(equivalenceTest)

}
