% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{getNboot}
\alias{getNboot}
\alias{getNboot.equivSDhtest}
\alias{getNboot.equivSDhtestList}
\alias{getNboot.AllEquivSDhtest}
\title{Acces to the number of effective bootstrap replicates}
\usage{
getNboot(x, ...)

\method{getNboot}{equivSDhtest}(x)

\method{getNboot}{equivSDhtestList}(x, simplify = TRUE)

\method{getNboot}{AllEquivSDhtest}(x, onto, GOLevel, listNames, simplify = TRUE)
}
\description{
Acces to the number of effective bootstrap replicates
}
\section{Methods (by class)}{
\itemize{
\item \code{equivSDhtest}: S3 method for class "equivSDhtest"

\item \code{equivSDhtestList}: S3 method for class "equivSDhtestList"

\item \code{AllEquivSDhtest}: S3 method for class "AllEquivSDhtest"
}}

