% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{getPvalue}
\alias{getPvalue}
\alias{getPvalue.equivSDhtest}
\alias{getPvalue.equivSDhtestList}
\alias{getPvalue.AllEquivSDhtest}
\title{Access to the p-value of one or more equivalence test results}
\usage{
getPvalue(x, ...)

\method{getPvalue}{equivSDhtest}(x, ...)

\method{getPvalue}{equivSDhtestList}(x, simplify = TRUE, ...)

\method{getPvalue}{AllEquivSDhtest}(x, onto, GOLevel, listNames, simplify = TRUE, ...)
}
\arguments{
\item{x}{an object of class "equivSDhtest" or "equivSDhtestList" or
"allEquivSDtest".}

\item{...}{Additional parameters.}

\item{simplify}{logical, if TRUE the result is simplified, e.g., returning a
vector instead
of a matrix.}

\item{onto}{character, a vector with one or more of "BP", "CC" or "MF",
ontologies to access.}

\item{GOLevel}{numeric or character, a vector with one or more GO levels to
access.
See the details section and the examples.}

\item{listNames}{character(2), the names of a pair of gene lists.}
}
\value{
When \code{x} is an object of class "equivSDhtest" (i.e., the result
of a single
equivalence test), the returned value is a single numeric value, the test
p-value.
For an object of class "equivSDhtestList" (i.e. all pairwise tests for a
set of gene lists), if \code{simplify = TRUE} (the default), the resulting
value is a vector
with the p-values in all those tests, or the symmetric matrix of all p-values
if \code{simplify = TRUE}. If \code{x} is an object of class "allEquivSDtest"
(i.e., the test iterated along GO ontologies and levels), the preceding
result is returned in the form of a list along the ontologies, levels and
pairs of gene lists specified by the arguments \code{onto, GOlevel} and
\code{listNames} (or all present in \code{x} for missing arguments).
}
\description{
Given objects representing the result(s) of one or more equivalence tests
(classes "equivSDhtest", "equivSDhtestList" or "allEquivSDtest", i.e., the
result of functions 'equivTestSorensen' and 'allEquivTestSorensen')
this function returns the p-values of the tests.
}
\details{
Argument \code{GOLevel} can be of class "character" or "numeric". In the
first case, the GO levels must be specified like \code{"level 6"} or
\code{c("level 4", "level 5", "level 6")}
In the second case ("numeric"), the GO levels must be specified like\code{6}
or \code{seq.int(4,6)}.
}
\section{Methods (by class)}{
\itemize{
\item \code{getPvalue(equivSDhtest)}: S3 method for class "equivSDhtest"

\item \code{getPvalue(equivSDhtestList)}: S3 method for class "equivSDhtestList"

\item \code{getPvalue(AllEquivSDhtest)}: S3 method for class "AllEquivSDhtest"

}}
\examples{
# Manually define a 2 x 2 enrichment contingency table
contTable <- as.table(matrix(c(127, 19, 159, 3018),
  nrow = 2,
  dimnames = list(
    "Enriched in List 1" = c(TRUE, FALSE),
    "Enriched in List 2" = c(TRUE, FALSE)
  )
))
contTable

# Compute the Sorensen equivalence test from the contingency table.
# The result is an object of class "equivSDhtest".
equivalenceTest <- equivTestSorensen(contTable)
equivalenceTest

# Extract the pvalue from the equivalence test object
getPvalue(equivalenceTest)

}
