% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getTable}
\alias{getTable}
\alias{getTable.equivSDhtest}
\alias{getTable.equivSDhtestList}
\alias{getTable.AllEquivSDhtest}
\title{Access to the contingency table of mutual enrichment of one or more
equivalence test results}
\usage{
getTable(x, ...)

\method{getTable}{equivSDhtest}(x, ...)

\method{getTable}{equivSDhtestList}(x, ...)

\method{getTable}{AllEquivSDhtest}(x, onto, GOLevel, listNames, ...)
}
\arguments{
\item{x}{an object of class "equivSDhtest" or "equivSDhtestList" or
"allEquivSDtest".}

\item{...}{Additional parameters.}

\item{onto}{character, a vector with one or more of "BP", "CC" or "MF",
ontologies to access.}

\item{GOLevel}{numeric or character, a vector with one or more GO levels to
access. See the details section and the examples.}

\item{listNames}{character(2), the names of a pair of gene lists.}
}
\value{
An object of class "table", the 2x2 enrichment contingeny table of
mutual enrichment in two gene lists, built to perform the equivalence test
based on the Sorensen-Dice dissimilarity.

When \code{x} is an object of class "equivSDhtest" (i.e., the result
of a single equivalence test), the returned value is an object of class
"table", the 2x2 enrichment contingeny table of mutual enrichment in two gene
lists, built to perform the equivalence test based on the Sorensen-Dice
dissimilarity.
For an object of class "equivSDhtestList" (i.e. all pairwise tests for a
set of gene lists), the resulting value is a list with all the tables built
in all those tests. If \code{x} is an object of class "allEquivSDtest"
(i.e., the test iterated along GO ontologies and levels), the preceding
result is returned as a list along the ontologies, levels and pairs of gene
lists specified by the arguments \code{onto, GOlevel} and
\code{listNames} (or all ontologies, levels or pairs of gene lists
present in \code{x} if one or more of these arguments are missing).
}
\description{
Given objects representing the result(s) of one or more equivalence tests
(classes "equivSDhtest", "equivSDhtestList" or "allEquivSDtest", i.e., the
result of functions 'equivTestSorensen' and 'allEquivTestSorensen')
this function returns the contingency tables from which the tests were
performed.
}
\details{
Argument \code{GOLevel} can be of class "character" or "numeric". In the
first case, the GO levels must be specified like \code{"level 6"} or
\code{c("level 4", "level 5", "level 6")} In the second case ("numeric"),
the GO levels must be specified like\code{6} or \code{4:6}.
}
\section{Methods (by class)}{
\itemize{
\item \code{getTable(equivSDhtest)}: S3 method for class "equivSDhtest"

\item \code{getTable(equivSDhtestList)}: S3 method for class "equivSDhtestList"

\item \code{getTable(AllEquivSDhtest)}: S3 method for class "AllEquivSDhtest"

}}
\examples{
# Manually define a 2 x 2 enrichment contingency table
contTable <- as.table(matrix(c(127, 19, 159, 3018),
  nrow = 2,
  dimnames = list(
    "Enriched in List 1" = c(TRUE, FALSE),
    "Enriched in List 2" = c(TRUE, FALSE)
  )
))
contTable

# Compute the Sorensen equivalence test from the contingency table.
# The result is an object of class "equivSDhtest".
equivalenceTest <- equivTestSorensen(contTable)
equivalenceTest

# Access the enrichment contingency table
getTable(equivalenceTest)

}
