% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTest.R
\name{GOIDsInLevel}
\alias{GOIDsInLevel}
\title{GOIDsInLevel}
\usage{
GOIDsInLevel(
  GOLev,
  onto,
  geneList1 = NULL,
  geneList2 = NULL,
  orgPackage = NULL,
  restricted = TRUE
)
}
\arguments{
\item{GOLev}{An integer}

\item{onto}{string describing the ontology. Belongs to c('BP', 'MF', 'CC', 'ANY')}

\item{geneList1}{character vector containing a FIRST gene list of entrez IDs}

\item{geneList2}{character vector containing a SECOND gene list of entrez IDs}

\item{orgPackage}{A string wih the name of the annotation package}

\item{restricted}{Boolean variable to decide how tabulation is performed.}
}
\description{
This function extends getGOLevel returning only GO identifiers appearing between GO ancestors of at least one GeneList
}
\examples{
 data(kidneyGeneLists)
 GOLev <- 3
 anOnto <- 'BP'
 gl1 <- kidneyGeneLists[[1]]
 gl2 <- kidneyGeneLists[[2]]
 orgPkg ='org.Hs.eg.db'
 GOIDSinLevel.1 <- GOIDsInLevel (GOLev=GOLev, onto = anOnto,  restricted=FALSE);
 length(GOIDSinLevel.1)
 GOIDSinLevel.2 <- GOIDsInLevel (GOLev=GOLev, onto = anOnto,  geneList1 = gl1,
 geneList2 = gl2, orgPackage =orgPkg)
 length(GOIDSinLevel.2)
}
