% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgu133plus2EntrezIDs}
\alias{hgu133plus2EntrezIDs}
\title{Entrez Identifiers for all human genes contained in the hgu133plus2.db package.}
\format{
An object of class \code{character} of length 22013.
}
\source{
\url{http://bioconductor.org}
}
\usage{
data(hgu133plus2EntrezIDs)
}
\description{
Entrez Identifiers for all human genes contained in the hgu133plus2.db package.
}
\keyword{datasets}
