% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanEntrezIDs}
\alias{humanEntrezIDs}
\title{Entrez Identifiers for all human genes contained in the org.Hs.eg.db package.}
\format{
An object of class \code{character} of length 61521.
}
\source{
\url{http://bioconductor.org}
}
\usage{
data(humanEntrezIDs)
}
\description{
Entrez Identifiers for all human genes contained in the org.Hs.eg.db package.
}
\keyword{datasets}
