% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kidneyEnrichedGOIDs}
\alias{kidneyEnrichedGOIDs}
\title{GOTerms identifiers obtained from an enrichment analysis performed on kidney Gene lists}
\format{
An object of class \code{list} of length 5.
}
\source{
\url{https://www.ualberta.ca/medicine/institutes-centres-groups/atagc/research/gene-lists}
}
\usage{
data(kidneyEnrichedGOIDs)
}
\description{
A dataset 5 character vectors representing lists of enriched GO Terms.
}
\keyword{datasets}
