% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kidneyGeneLists}
\alias{kidneyGeneLists}
\title{Gene lists derived from studies on kidney transplantation rejection}
\format{
An object of class \code{list} of length 5.
}
\source{
\url{https://www.ualberta.ca/medicine/institutes-centres-groups/atagc/research/gene-lists}
}
\usage{
data(kidneyGeneLists)
}
\description{
A dataset 5 character vectors representing 5 gene lists.
}
\keyword{datasets}
