% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbtGeneLists.R
\docType{data}
\name{pbtGeneLists}
\alias{pbtGeneLists}
\title{5 gene lists possibly related with kidney transplant rejection}
\format{
An object of class "list" of length 5. Each one of its elements is a "character" vector
of gene identifiers.
}
\source{
\url{https://www.ualberta.ca/medicine/institutes-centres-groups/atagc/research/gene-lists.html}
}
\usage{
data(pbtGeneLists)
}
\description{
An object of class "list" of length 5. A non up-to-date subset of the University of Alberta pathogenesis-based
transcripts sets (PBTs) that were generated by using Affymetrix Microarrays. Take them just as an illustrative
example.
}
\keyword{datasets}
