% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTest.R
\name{stdTest4GOIDs}
\alias{stdTest4GOIDs}
\title{StandardTest4GOIDs}
\usage{
stdTest4GOIDs(GO1, GO2, onto, GOLevel, listNames = NULL)
}
\arguments{
\item{GO1}{character vector containing a FIRST list of GO identifiers}

\item{GO2}{character vector containing a SECOND gene list GO identifiers}

\item{onto}{string describing the ontology. Belongs to c('BP', 'MF', 'CC', 'ANY')}

\item{listNames}{character vector with names of the genelists that generated the enriched GOIDs}

\item{GOLev}{An integer}
}
\description{
This function performs a "standard test" of equivalence between two lists of enriched GOTerms
The lists are intended to have been obtained from enrichment analyses performed on two gene lists
Given that this version of the test does not provide gene lists only "unrestricted" crosstabulation can be performed
That is, all terms in the given GO level are used to build the table, even if they are not found in the geneLists's annotations.
}
\examples{
 GOIDs2 <-kidneyEnrichedGOIDs[[2]]
 data(kidneyEnrichedGOIDs)
 GOIDs1 <-kidneyEnrichedGOIDs[[1]]
 names4lists<- names(kidneyEnrichedGOIDs)[1:2]
 anOnto <- 'BP'
 GOLev<- 3
 testedFromGOIds <- stdTest4GOIDs (GO1 = GOIDs1, GO2 = GOIDs2, onto = anOnto,
 GOLevel =GOLev, listNames=names4lists)
}
