% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgrade}
\alias{upgrade}
\alias{upgrade.equivSDhtest}
\alias{upgrade.equivSDhtestList}
\alias{upgrade.AllEquivSDhtest}
\title{Update the result of a Sorensen-Dice equivalence test.}
\usage{
upgrade(x, ...)

\method{upgrade}{equivSDhtest}(x, ...)

\method{upgrade}{equivSDhtestList}(x, ...)

\method{upgrade}{AllEquivSDhtest}(x, ...)
}
\arguments{
\item{x}{an object of class "equivSDhtest", "equivSDhtestList" or
"AllEquivSDhtest".}

\item{...}{any valid parameters for function "equivTestSorensen" for its
interface "table", to recompute the test(s) according to these parameters.}
}
\value{
An object of the same class than \code{x}.
}
\description{
Recompute the test (or tests) from an object of class "equivSDhtest",
"equivSDhtestList" or "AllEquivSDhtest" (i.e.,the output of functions
"equivTestSorensen" or "allEquivTestSorensen").
Using the same table or tables of enrichment frequencies in 'x', obtain again
the result of the equivalence test for new values of any of the parameters
\code{d0} or \code{conf.level} or \code{boot} or \code{nboot} or
\code{check.table}.
}
\section{Methods (by class)}{
\itemize{
\item \code{upgrade(equivSDhtest)}: S3 method for class "equivSDhtest"

\item \code{upgrade(equivSDhtestList)}: S3 method for class "equivSDhtestList"

\item \code{upgrade(AllEquivSDhtest)}: S3 method for class "allEquivSDhtest"

}}
\examples{
# Manually define a 2 x 2 enrichment contingency table
contTable <- as.table(matrix(c(127, 19, 159, 3018),
  nrow = 2,
  dimnames = list(
    "Enriched in List 1" = c(TRUE, FALSE),
    "Enriched in List 2" = c(TRUE, FALSE)
  )
))
contTable

# Compute the Sorensen equivalence test from the contingency table.
# The result is an object of class "equivSDhtest".
equivalenceTest <- equivTestSorensen(contTable)
equivalenceTest

# Access the equivalence test results.
upgrade(equivalenceTest)

}
