% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEnrichTable.R
\name{allBuildEnrichTable}
\alias{allBuildEnrichTable}
\title{Iterate \code{buildEnrichTable} along the specified GO ontologies and GO levels}
\usage{
allBuildEnrichTable(
  x,
  check.table = TRUE,
  ontos = c("BP", "CC", "MF"),
  GOLevels = seq.int(3, 10),
  trace = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class "list". Each of its elements must be a "character" vector of gene
identifiers. Then all pairwise contingency tables of joint enrichment are built between
these gene lists, iterating the process for all specified GO ontologies and GO levels.}

\item{check.table}{Boolean. If TRUE (default), all resulting tables are checked by means
of function \code{nice2x2Table}.}

\item{ontos}{"character", GO ontologies to analyse. Defaults to \code{c("BP", "CC", "MF")}.}

\item{GOLevels}{"integer", GO levels to analyse inside each of these GO ontologies.}

\item{trace}{Logical. If TRUE (default), the (usually very time consuming) process of function
\code{allBuildEnrichTable} is traced along the specified GO ontologies and levels.}

\item{...}{extra parameters for function \code{buildEnrichTable}.}
}
\value{
An object of class "allTableList". It is a list with as many components as GO ontologies have been
analysed.
Each of these elements is itself a list with as many components as GO levels have been analised.
Finally, the elements of these lists are objects as generated by \code{buildEnrichTable.list},
i.e., objects of class "tableList" containing all pairwise contingency tables of mutual enrichmen
between the gene lists in argument \code{x}.
}
\description{
Iterate \code{buildEnrichTable} along the specified GO ontologies and GO levels
}
\examples{
# This example is extremely time consuming, it scans two GO ontologies and three
# GO levels inside them to obtain the contingency tables of mutual enrichment.
# Gene universe:
# data(humanEntrezIDs)
# Gene lists to be explored for enrichment:
# data(pbtGeneLists)
# allBuildEnrichTable(pbtGeneLists,
#                     geneUniverse = humanEntrezIDs, orgPackg = "org.Hs.eg.db",
#                     ontos = c("MF", "BP"), GOLevels = seq.int(4,6))

}
