% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivtestsorensen.R
\name{allEquivTestSorensen}
\alias{allEquivTestSorensen}
\title{Iterate \code{equivTestSorensen} along the specified GO ontologies and GO levels}
\usage{
allEquivTestSorensen(
  x,
  d0 = 1/(1 + 1.25),
  conf.level = 0.95,
  boot = FALSE,
  nboot = 10000,
  check.table = TRUE,
  ontos = c("BP", "CC", "MF"),
  GOLevels = 3:10,
  ...
)
}
\arguments{
\item{x}{object of class "list". Each of its elements must be a "character" vector of gene
identifiers. Then all pairwise equivalence tests are performed between these gene lists,
iterating the process for all specified GO ontologies and GO levels.}

\item{d0}{equivalence threshold for the population Sorensen-Dice dissimilarity, d. The null hypothesis
states that d >= d0, and the alternative that d < d0.}

\item{conf.level}{confidence level of the one-sided confidence interval, a value between 0 and 1.}

\item{ontos}{"character", GO ontologies to analyse. Defaults to \code{c("BP", "CC", "MF")}.}

\item{GOLevels}{"integer", GO levels to analyse inside each of these GO ontologies.}

\item{...}{extra parameters for function \code{crossTabGOIDs4GeneLists} in package \code{equivStandardTest}.}
}
\value{
An object of class "AllEquivSDhtest". It is a list with as many components as GO ontologies have been analysed.
Each of these elements is itself a list with as many components as GO levels have been analized.
Finally, the elements of these lists are objects as generated by \code{equivTestSorensen.list},
i.e., objects of class "equivSDhtestList" containing all pairwise comparisons between the gene lists
in argument \code{x}.
}
\description{
Iterate \code{equivTestSorensen} along the specified GO ontologies and GO levels
}
\examples{
library(equivStandardTest)
# Gene universe:
data(humanEntrezIDs)
# Gene lists to be explored for enrichment:
data(pbtGeneLists)
allEquivTestSorensen(pbtGeneLists,
                    geneUniverse = humanEntrezIDs, orgPackg = "org.Hs.eg.db",
                    ontos = c("MF", "BP"), GOLevels = 4:6)

}
