% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeTable.R
\name{completeTable}
\alias{completeTable}
\title{Reformats and completes (if necessary) a 2x2 enrichment contingency table for its appropriate use in package goSorensen. It is internally used by function 'buildEnrichTable'..}
\usage{
completeTable(x, listNames)
}
\arguments{
\item{x}{an object of class "table".}

\item{listNames}{a character(2) with the gene lists names.
enrichment frequencies.}
}
\value{
a complete contingency table to use in package goSorensen.
}
\description{
Reformats and completes (if necessary) a 2x2 enrichment contingency table for its appropriate use in package goSorensen. It is internally used by function 'buildEnrichTable'..
}
