% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTest.R
\name{crossTabGOIDsUnrestricted}
\alias{crossTabGOIDsUnrestricted}
\title{crossTabGOIDsUnrestricted}
\usage{
crossTabGOIDsUnrestricted(GO1, GO2, onto, GOLev, listNames = NULL)
}
\arguments{
\item{GO1}{character vector containing a FIRST list of GO identifiers}

\item{GO2}{character vector containing a SECOND gene list GO identifiers}

\item{onto}{string describing the ontology. Belongs to c('BP', 'MF', 'CC', 'ANY')}

\item{GOLev}{An integer}

\item{listNames}{character vector with names of the genelists that generated the
enriched GOIDs}
}
\description{
This function performs a crosstabulation between two lists of enriched GOTerms
The lists are intended to have been obtained from enrichment analyses performed
on two gene lists
}
\examples{
 data(kidneyEnrichedGOIDs)
 GOIDs1 <-kidneyEnrichedGOIDs[[1]]
 GOIDs2 <-kidneyEnrichedGOIDs[[2]]
 names4lists<- names(kidneyEnrichedGOIDs)[1:2]
 anOnto <- 'BP'
 GOLev<- 3
 crossTabbedGOIds <- crossTabGOIDsUnrestricted (GO1 = GOIDs1, GO2 = GOIDs2, onto = anOnto,
 GOLev =GOLev, listNames=names4lists)
}
