% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{getDissimilarity}
\alias{getDissimilarity}
\alias{getDissimilarity.equivSDhtest}
\alias{getDissimilarity.equivSDhtestList}
\alias{getDissimilarity.AllEquivSDhtest}
\title{Acces to the Sorensen-Dice dissimilarity}
\usage{
getDissimilarity(x, ...)

\method{getDissimilarity}{equivSDhtest}(x)

\method{getDissimilarity}{equivSDhtestList}(x, simplify = TRUE)

\method{getDissimilarity}{AllEquivSDhtest}(x, onto, GOLevel, listNames, simplify = TRUE)
}
\description{
Acces to the Sorensen-Dice dissimilarity
}
\section{Methods (by class)}{
\itemize{
\item \code{equivSDhtest}: S3 method for class "equivSDhtest"

\item \code{equivSDhtestList}: S3 method for class "equivSDhtestList"

\item \code{AllEquivSDhtest}: S3 method for class "AllEquivSDhtest"
}}

