% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNboot.R
\name{getNboot}
\alias{getNboot}
\alias{getNboot.equivSDhtest}
\alias{getNboot.equivSDhtestList}
\alias{getNboot.AllEquivSDhtest}
\title{Access to the number of initially planned bootstrap replicates in one or more equivalence test
results (only in their bootstrap version)}
\usage{
getNboot(x, ...)

\method{getNboot}{equivSDhtest}(x, ...)

\method{getNboot}{equivSDhtestList}(x, simplify = TRUE, ...)

\method{getNboot}{AllEquivSDhtest}(x, onto, GOLevel, listNames, simplify = TRUE, ...)
}
\arguments{
\item{x}{an object of class "equivSDhtest" or "equivSDhtestList" or "allEquivSDtest".}

\item{...}{Additional parameters.}

\item{simplify}{logical, if TRUE the result is simplified, e.g., returning a vector instead
of a matrix.}

\item{onto}{character, a vector with one or more of "BP", "CC" or "MF", ontologies to access.}

\item{GOLevel}{numeric or character, a vector with one or more GO levels to access.
See the details section and the examples.}

\item{listNames}{character(2), the names of a pair of gene lists.}
}
\value{
When \code{x} is an object of class "equivSDhtest" (i.e., the result of a single
equivalence test), the returned value is a single numeric value, the number of initially
planned bootstrap replicates, or \code{NA} if bootstrapping has not been performed.
For an object of class "equivSDhtestList" (i.e. all pairwise tests for a
set of gene lists), if \code{simplify = TRUE} (the default), the resulting value is a vector
with the number of initially bootstrap replicates in all those tests, or the symmetric matrix
of all these values if \code{simplify = TRUE}.
If \code{x} is an object of class "allEquivSDtest"
(i.e., the test iterated along GO ontologies and levels), the preceding result is returned in
the form of a list along the ontologies, levels and pairs of gene lists specified by the arguments
\code{onto, GOlevel} and \code{listNames} (or all present in \code{x} for missing arguments).
}
\description{
Given objects representing the result(s) of one or more equivalence tests
(classes "equivSDhtest", "equivSDhtestList" or "allEquivSDtest", i.e., the
result of functions 'equivTestSorensen' and 'allEquivTestSorensen' with the parameter
boot = TRUE), this function returns the number of initially planned bootstrap replicates
in these equivalence tests, which may be greater than the number of finally effective or
valid bootstrap replicates. See the details section for more information on this.
}
\details{
In the bootstrap version of the equivalence test, resampling is performed generating new
bootstrap contingency tables from a multinomial distribution based on the "real", observed,
frequencies of mutual enrichment.
In some bootstrap iterations, the generated contingency table of mutual enrichment
may have very low frequencies of enrichment, which makes it unable for Sorensen-Dice
computations.
Then, the number of effective bootstrap resamples may be lower than those initially planned.
To get the number of effective bootstrap resamples use function \code{getEffNboot}.

Argument \code{GOLevel} can be of class "character" or "numeric". In the first case, the GO
levels must be specified like \code{"level 6"} or \code{c("level 4", "level 5", "level 6")}
In the second case ("numeric"), the GO levels must be specified like\code{6} or \code{seq.int(4,6)}.
}
\section{Methods (by class)}{
\itemize{
\item \code{getNboot(equivSDhtest)}: S3 method for class "equivSDhtest"

\item \code{getNboot(equivSDhtestList)}: S3 method for class "equivSDhtestList"

\item \code{getNboot(AllEquivSDhtest)}: S3 method for class "AllEquivSDhtest"

}}
\examples{
# Dataset 'allOncoGeneLists' contains the result of the equivalence test between gene lists
# 'waldman' and 'atlas', at level 4 of the BP ontology:
data(waldman_atlas.BP.4)
waldman_atlas.BP.4
class(waldman_atlas.BP.4)
# This may correspond to the result of code like:
# waldman_atlas.BP.4 <- equivTestSorensen(
#   allOncoGeneLists[["waldman"]], allOncoGeneLists[["atlas"]],
#   geneUniverse = humanEntrezIDs, orgPackg = "org.Hs.eg.db",
#   onto = "BP", GOLevel = 4, listNames = c("waldman", "atlas"))
#
# (But results may vary according to GO updating)

# Not a bootstrap test, first upgrade to a bootstrap test:
boot.waldman_atlas.BP.4 <- upgrade(waldman_atlas.BP.4, boot = TRUE)

getNboot(waldman_atlas.BP.4)
getNboot(boot.waldman_atlas.BP.4)

# All pairwise equivalence tests at level 4 of the BP ontology
data(BP.4)
?BP.4
class(BP.4)
# This may correspond to a call like:
# BP.4 <- equivTestSorensen(allOncoGeneLists,
#                           geneUniverse = humanEntrezIDs, orgPackg = "org.Hs.eg.db",
#                           onto = "BP", GOLevel = 4)
boot.BP.4 <- upgrade(BP.4, boot = TRUE)
getNboot(BP.4)
getNboot(boot.BP.4)
getNboot(boot.BP.4, simplify = FALSE)

# Bootstrap equivalence test iterated over all GO ontologies and levels 3 to 10.
# data(cancerEquivSorensen)
# ?cancerEquivSorensen
# class(cancerEquivSorensen)
# This may correspond to code like:
# (By default, the tests are iterated over all GO ontologies and for levels 3 to 10)
# cancerEquivSorensen <- allEquivTestSorensen(allOncoGeneLists,
#                                             geneUniverse = humanEntrezIDs,
#                                             orgPackg = "org.Hs.eg.db",
#                                             boot = TRUE)
# boot.cancerEquivSorensen <- upgrade(cancerEquivSorensen, boot = TRUE)
# All numbers of bootstrap replicates:
# getNboot(boot.cancerEquivSorensen)
# getNboot(boot.cancerEquivSorensen, simplify = FALSE)

# Number of bootstrap replicates for specific GO ontologies, levels or pairs of gene lists:
# getNboot(boot.cancerEquivSorensen, GOLevel = "level 6")
# getNboot(boot.cancerEquivSorensen, GOLevel = 6)
# getNboot(boot.cancerEquivSorensen, GOLevel = seq.int(4,6))
# getNboot(boot.cancerEquivSorensen, GOLevel = "level 6", simplify = FALSE)
# getNboot(boot.cancerEquivSorensen, GOLevel = "level 6", listNames = c("waldman", "sanger"))
# getNboot(boot.cancerEquivSorensen, GOLevel = seq.int(4,6), onto = "BP")
# getNboot(boot.cancerEquivSorensen, GOLevel = seq.int(4,6), onto = "BP", simplify = FALSE)
# getNboot(boot.cancerEquivSorensen, GOLevel = "level 6", onto = "BP",
#          listNames = c("waldman", "sanger"))
# getNboot(boot.cancerEquivSorensen$BP$`level 4`)

}
\seealso{
\code{\link{getEffNboot}}
}
