% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{getPvalue}
\alias{getPvalue}
\alias{getPvalue.equivSDhtest}
\alias{getPvalue.equivSDhtestList}
\alias{getPvalue.AllEquivSDhtest}
\title{Acces to Sorensen-Dice asymptotic test p-value}
\usage{
getPvalue(x, ...)

\method{getPvalue}{equivSDhtest}(x)

\method{getPvalue}{equivSDhtestList}(x, simplify = TRUE)

\method{getPvalue}{AllEquivSDhtest}(x, onto, GOLevel, listNames, simplify = TRUE)
}
\description{
Acces to Sorensen-Dice asymptotic test p-value
}
\section{Methods (by class)}{
\itemize{
\item \code{equivSDhtest}: S3 method for class "equivSDhtest"

\item \code{equivSDhtestList}: S3 method for class "equivSDhtestList"

\item \code{AllEquivSDhtest}: S3 method for class "AllEquivSDhtest"
}}

