% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{getSE}
\alias{getSE}
\alias{getSE.equivSDhtest}
\alias{getSE.equivSDhtestList}
\alias{getSE.AllEquivSDhtest}
\title{Acces to Sorensen-Dice asymptotic test standard error of the dissimilarity estimate}
\usage{
getSE(x, ...)

\method{getSE}{equivSDhtest}(x)

\method{getSE}{equivSDhtestList}(x, simplify = TRUE)

\method{getSE}{AllEquivSDhtest}(x, onto, GOLevel, listNames, simplify = TRUE)
}
\description{
Acces to Sorensen-Dice asymptotic test standard error of the dissimilarity estimate
}
\section{Methods (by class)}{
\itemize{
\item \code{equivSDhtest}: S3 method for class "equivSDhtest"

\item \code{equivSDhtestList}: S3 method for class "equivSDhtestList"

\item \code{AllEquivSDhtest}: S3 method for class "AllEquivSDhtest"
}}

