% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getTable}
\alias{getTable}
\alias{getTable.equivSDhtest}
\alias{getTable.equivSDhtestList}
\alias{getTable.AllEquivSDhtest}
\title{Access to the contingency table of mutual enrichment of one or more equivalence test results}
\usage{
getTable(x, ...)

\method{getTable}{equivSDhtest}(x, ...)

\method{getTable}{equivSDhtestList}(x, ...)

\method{getTable}{AllEquivSDhtest}(x, onto, GOLevel, listNames, ...)
}
\arguments{
\item{x}{an object of class "equivSDhtest" or "equivSDhtestList" or "allEquivSDtest".}

\item{...}{Additional parameters.}

\item{onto}{character, a vector with one or more of "BP", "CC" or "MF", ontologies to access.}

\item{GOLevel}{numeric, a vector with one or more GO levels to access.}

\item{listNames}{character(2), the names of a pair of gene lists.}
}
\description{
Given objects representing the result(s) of one or more equivalence tests
(classes "equivSDhtest", "equivSDhtestList" or "allEquivSDtest", i.e., the
result of functions 'equivTestSorensen' and 'allEquivTestSorensen')
this function returns the contingency tables from which the tests were performed.
}
\section{Methods (by class)}{
\itemize{
\item \code{equivSDhtest}: S3 method for class "equivSDhtest"

\item \code{equivSDhtestList}: S3 method for class "equivSDhtestList"

\item \code{AllEquivSDhtest}: S3 method for class "AllEquivSDhtest"
}}

\examples{
# Dataset 'allOncoGeneLists' contains the result of the equivalence test between gene lists
# 'waldman' and 'atlas', at level 4 of the BP ontology:
waldman_atlas.BP.4
class(waldman_atlas.BP.4)
# This may correspond to the result of code like:
# waldman_atlas.BP.4 <- equivTestSorensen(
#   allOncoGeneLists[["waldman"]], allOncoGeneLists[["atlas"]],
#   geneUniverse = humanEntrezIDs, orgPackg = "org.Hs.eg.db",
#   onto = "BP", GOLevel = 4, listNames = c("waldman", "atlas"))
# (But results may vary according to GO updating)
getTable(waldman_atlas.BP.4)

# All pairwise equivalence tests at level 4 of the BP ontology
BP.4
class(BP.4)
# This may correspond to a call like:
# BP.4 <- equivTestSorensen(allOncoGeneLists,
#                           geneUniverse = humanEntrezIDs, orgPackg = "org.Hs.eg.db",
#                           onto = "BP", GOLevel = 4)
getTable(BP.4)

# Equivalence test iterated over all GO ontologies and levels 3 to 10:
cancerEquivSorensen
class(cancerEquivSorensen)
# This may correspond to code like:
# cancerEquivSorensen <- allEquivTestSorensen(allOncoGeneLists,
#                                             geneUniverse = humanEntrezIDs,
#                                             orgPackg = "org.Hs.eg.db")
# (By default, the tests are iterated over all GO ontologies and for levels 3 to 10)
# All 2x2 contingecy tables of joint enrichment:
getTable(cancerEquivSorensen)
# Contingency tables only for some GO ontologies, levels or pairs of gene lists:
getTable(cancerEquivSorensen, GOLevel = "level 6")
getTable(cancerEquivSorensen, GOLevel = "level 6", listNames = c("waldman", "sanger"))
getTable(cancerEquivSorensen, GOLevel = "level 6", onto = "BP")
getTable(cancerEquivSorensen, GOLevel = "level 6", onto = "BP",
         listNames = c("waldman", "sanger"))


}
