% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{getUpper}
\alias{getUpper}
\alias{getUpper.equivSDhtest}
\alias{getUpper.equivSDhtestList}
\alias{getUpper.AllEquivSDhtest}
\title{Acces to Sorensen-Dice asymptotic test confidence interval upper limit}
\usage{
getUpper(x, ...)

\method{getUpper}{equivSDhtest}(x)

\method{getUpper}{equivSDhtestList}(x, simplify = TRUE)

\method{getUpper}{AllEquivSDhtest}(x, onto, GOLevel, listNames, simplify = TRUE)
}
\description{
Acces to Sorensen-Dice asymptotic test confidence interval upper limit
}
\section{Methods (by class)}{
\itemize{
\item \code{equivSDhtest}: S3 method for class "equivSDhtest"

\item \code{equivSDhtestList}: S3 method for class "equivSDhtestList"

\item \code{AllEquivSDhtest}: S3 method for class "AllEquivSDhtest"
}}

