% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice2x2Table.R
\name{nice2x2Table}
\alias{nice2x2Table}
\alias{nice2x2Table.table}
\alias{nice2x2Table.matrix}
\alias{nice2x2Table.numeric}
\title{Checks for validity data representing an enrichment contingency table generated from two gene lists}
\usage{
nice2x2Table(x)

\method{nice2x2Table}{table}(x)

\method{nice2x2Table}{matrix}(x)

\method{nice2x2Table}{numeric}(x)
}
\arguments{
\item{x}{either an object of class "table", "matrix" or "numeric".}
}
\value{
boolean, TRUE if \code{x} nicely represents a 2x2 contingency table
interpretable as the cross-tabulation of the enriched GO items in two gene lists:
"Number of enriched items in list 1 (TRUE, FALSE)" x "Number of enriched items in
list 2 (TRUE, FALSE)". In this function, "nicely representing a 2x2 contingency table"
is interpreted in terms of computing the Sorensen-Dice dissimilarity and associated
statistics. In these computations, the double negatives n00 are ignored.
The result is FALSE otherwise and warnings are issued.
}
\description{
Checks for validity data representing an enrichment contingency table generated from two gene lists
}
\details{
In the "table" and "matrix" interfaces, the input parameter \code{x} must correspond
to a two-dimensional array:\tabular{ll}{
   n_11 \tab n_10 \cr
   n_01 \tab n_00 \cr
}


These values are interpreted (always in this order) as n11: number of GO items enriched in both lists,
n01: GO items enriched in the second list but not in the first one, n10: items not enriched in the second
list but enriched in the first one and double negatives, n00.
The double negatives n00 are ignored in many computations concerning the Sorensen-Dice index.

In the "numeric" interface, the input \code{x} must correspond to a numeric of length
3 or more, in the same order as before.
}
\section{Methods (by class)}{
\itemize{
\item \code{table}: S3 method for class "table"

\item \code{matrix}: S3 method for class "matrix"

\item \code{numeric}: S3 method for class "numeric"
}}

\examples{
conti <- as.table(matrix(c(27, 36, 12, 501, 43, 15, 0, 0, 0), nrow = 3, ncol = 3,
                         dimnames = list(c("a1","a2","a3"),
                                         c("b1", "b2","b3"))))
nice2x2Table(conti)
conti2 <- conti[1,1:min(2,ncol(conti)), drop = FALSE]
conti2
nice2x2Table(conti2)

conti3 <- matrix(c(12, 210), ncol = 2, nrow = 1)
conti3
nice2x2Table(conti3)

conti4 <- c(32, 21, 81, 1439)
nice2x2Table(conti4)
conti4.mat <- matrix(conti4, nrow = 2)
conti4.mat
conti5 <- c(32, 21, 81)
nice2x2Table(conti5)


}
