% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesorensen.R
\name{seSorensen}
\alias{seSorensen}
\alias{seSorensen.table}
\alias{seSorensen.matrix}
\alias{seSorensen.numeric}
\alias{seSorensen.character}
\alias{seSorensen.list}
\alias{seSorensen.tableList}
\title{Standard error of the sample Sorensen-Dice dissimilarity, asymptotic approach}
\usage{
seSorensen(x, ...)

\method{seSorensen}{table}(x, check.table = TRUE, ...)

\method{seSorensen}{matrix}(x, check.table = TRUE, ...)

\method{seSorensen}{numeric}(x, check.table = TRUE, ...)

\method{seSorensen}{character}(x, y, check.table = TRUE, ...)

\method{seSorensen}{list}(x, check.table = TRUE, ...)

\method{seSorensen}{tableList}(x, check.table = TRUE, ...)
}
\arguments{
\item{x}{either an object of class "table", "matrix" or "numeric" representing a 2x2 contingency table,
or a "character" (a set of gene identifiers) or "list" or "tableList" object.
See the details section for more information.}

\item{...}{extra parameters for function \code{buildEnrichTable}.}

\item{check.table}{Boolean. If TRUE (default), argument \code{x} is checked to adequately
represent a 2x2 contingency table. This checking is performed by means of function
\code{nice2x2Table}.}

\item{y}{an object of class "character" representing a vector of gene identifiers (e.g., ENTREZ).}
}
\value{
In the "table", "matrix", "numeric" and "character" interfaces, the value of the standard error of the
Sorensen-Dice dissimilarity estimate.
In the "list" and "tableList" interfaces, the symmetric matrix of all
standard error dissimilarity estimates.
}
\description{
Standard error of the sample Sorensen-Dice dissimilarity, asymptotic approach
}
\details{
This function computes the standard error estimate of the sample Sorensen-Dice dissimilarity,
given a 2x2 arrangement of frequencies (either implemented as a "table", a "matrix"
or a "numeric" object):
\tabular{rr}{
\eqn{n_{11}} \tab \eqn{n_{10}} \cr
\eqn{n_{01}} \tab \eqn{n_{00}},
}

The subindex '11' corresponds to those
GO terms enriched in both lists, '01' to terms enriched in the second list but not in the first one,
'10' to terms enriched in the first list but not enriched in the second one and '00' corresponds
to those GO terms non enriched in both gene lists, i.e., to the double negatives, a value which
is ignored in the computations.

In the "numeric" interface, if \code{length(x) >= 3}, the values are interpreted
as
\eqn{(n_{11}, n_{01}, n_{10})}{%
(n_11, n_01, n_10)}, always in this order.

If \code{x} is an object of class "character", then \code{x} (and \code{y}) must represent
two "character" vectors of valid gene identifiers (e.g., ENTREZ).
Then the standard error for the dissimilarity between lists \code{x} and \code{y} is computed,
after internally summarizing them as a 2x2 contingency table of joint enrichment.
This last operation is performed by function \code{\link{buildEnrichTable}} and "valid gene
identifiers (e.g., ENTREZ)" stands for the coherency of these gene identifiers with the arguments
\code{geneUniverse} and \code{orgPackg} of \code{buildEnrichTable}, passed by the ellipsis
argument \code{...} in \code{seSorensen}.

In the "list" interface, the argument must be a list of "character" vectors, each one
representing a gene list (character identifiers). Then, all pairwise standard errors
of the dissimilarity between these gene lists are computed.

If \code{x} is an object of class "tableList", the standard error of the Sorensen-Dice dissimilarity
estimate is computed over each one of these tables.
Given k gene lists (i.e. "character" vectors of gene identifiers) l1, l2, ..., lk,
an object of class "tableList" (typically constructed by a call to function
\code{\link{buildEnrichTable}}) is a list of lists of
contingency tables t(i,j) generated from each pair of gene lists i and j, with the
following structure:

$l2

$l2$l1$t(2,1)

$l3

$l3$l1$t(3,1), $l3$l2$t(3,2)

...

$lk

$lk$l1$t(k,1), $lk$l2$t(k,2), ..., $lk$l(k-1)t(k,k-1)
}
\section{Methods (by class)}{
\itemize{
\item \code{seSorensen(table)}: S3 method for class "table"

\item \code{seSorensen(matrix)}: S3 method for class "matrix"

\item \code{seSorensen(numeric)}: S3 method for class "numeric"

\item \code{seSorensen(character)}: S3 method for class "character"

\item \code{seSorensen(list)}: S3 method for class "list"

\item \code{seSorensen(tableList)}: S3 method for class "tableList"

}}
\examples{
# Gene lists 'atlas' and 'sanger' in 'allOncoGeneLists' dataset. Table of joint enrichment
# of GO terms in ontology BP at level 4.
data(cont_atlas.sanger_BP4)
cont_atlas.sanger_BP4
dSorensen(cont_atlas.sanger_BP4)
seSorensen(cont_atlas.sanger_BP4)

# Contingency table as a numeric vector:
seSorensen(c(56, 1, 30, 47))
seSorensen(c(56, 1, 30))

# (These examples may be considerably time consuming due to many enrichment
# tests to build the contingency tables of mutual enrichment)
# data(allOncoGeneLists)
# ?allOncoGeneLists

# Standard error of the sample Sorensen-Dice dissimilarity, directly from
# two gene lists, from scratch:
# seSorensen(allOncoGeneLists$atlas, allOncoGeneLists$sanger,
#            onto = "BP", GOLevel = 3,
#            geneUniverse = humanEntrezIDs, orgPackg = "org.Hs.eg.db")
# Essentially, the above code makes the same as:
# cont_atlas.sanger_BP4 <- buildEnrichTable(allOncoGeneLists$atlas, allOncoGeneLists$sanger,
#                                     onto = "BP", GOLevel = 4,
#                                     geneUniverse = humanEntrezIDs, orgPackg = "org.Hs.eg.db")
# cont_atlas.sanger_BP4
# seSorensen(cont_atlas.sanger_BP4)

# All pairwise standard errors (quite time consuming):
# seSorensen(allOncoGeneLists,
#            onto = "BP", GOLevel = 4,
#            geneUniverse = humanEntrezIDs, orgPackg = "org.Hs.eg.db")
}
\seealso{
\code{\link{buildEnrichTable}} for constructing contingency tables of mutual
enrichment,
\code{\link{nice2x2Table}} for checking the validity of enrichment contingency tables,
\code{\link{dSorensen}} for computing the Sorensen-Dice dissimilarity,
\code{\link{duppSorensen}} for the upper limit of a one-sided confidence interval
of the dissimilarity, \code{\link{equivTestSorensen}} for an equivalence test.
}
