% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesorensen.R
\name{seSorensen}
\alias{seSorensen}
\alias{seSorensen.table}
\alias{seSorensen.matrix}
\alias{seSorensen.numeric}
\alias{seSorensen.character}
\alias{seSorensen.list}
\title{Standard error of the sample Sorensen-Dice dissimilarity, asymptotic approach}
\usage{
seSorensen(x, ...)

\method{seSorensen}{table}(x, listNames = NULL, check.table = TRUE, ...)

\method{seSorensen}{matrix}(x, listNames = NULL, check.table = TRUE, ...)

\method{seSorensen}{numeric}(x, listNames = NULL, check.table = TRUE, ...)

\method{seSorensen}{character}(
  x,
  y,
  listNames = c("gene.list1", "gene.list2"),
  check.table = TRUE,
  ...
)

\method{seSorensen}{list}(x, check.table = TRUE, ...)
}
\arguments{
\item{x}{either an object of class "table", "matrix" or "numeric" representing a 2x2 contingency table,
or a "character" (a set of gene identifiers) or "list" object. See the details section for more information.}

\item{...}{extra parameters for function \code{buildEnrichTable}.}

\item{listNames}{character(2), names of both gene lists being compared (only for the character interface)}

\item{check.table}{Boolean. If TRUE (default), argument \code{x} is checked to adequately
represent a 2x2 contingency table. This checking is performed by means of function
\code{nice2x2Table}.}

\item{y}{an object of class "character" representing a vector of gene identifiers.}
}
\value{
In the "table", "matrix", "numeric" and "character" interfaces, the value of the standard error of the
Sorensen-Dice dissimilarity estimate. In the "list" interface, the symmetric matrix of all pairwise standard error
dissimilarity estimates.
}
\description{
Standard error of the sample Sorensen-Dice dissimilarity, asymptotic approach
}
\details{
This function computes the standard error estimate of the sample Sorensen-Dice dissimilarity,
given a 2x2 arrangement of frequencies (either implemented as a "table", a "matrix"
or a "numeric" object):
\deqn{
 \tabular{rr}{
  n_{11} \tab n_{10} \cr
  n_{01} \tab n_{00},
 }
}{}

\tabular{rr}{
n_11 \tab n_10 \cr
n_01 \tab n_00,
}

The subindex '11' corresponds to those
GO items enriched in both lists, '01' to items enriched in the second list but not in the first one,
'10' to items enriched in the first list but not enriched in the second one and '00' corresponds
to those GO items non enriched in both gene lists, i.e., to the double negatives, a value which
is ignored in the computations.

In the "numeric" interface, if \code{length(x) >= 3}, the values are interpreted
as
\eqn{(n_{11}, n_{01}, n_{10})}{%
(n_11, n_01, n_10)}, always in this order.

If \code{x} is an object of class "character", it must represent a list of gene identifiers. Then the
standard error for the dissimilarity between lists \code{x} and \code{y} is computed, after summarizing
these gene lists as a 2x2 contingency table of joint enrichment.

In the "list" interface, the argument must be a list of "character" vectors, each one
representing a gene list (character identifiers). Then, all pairwise standard errors
of the dissimilarity between these gene lists are computed.
}
\section{Methods (by class)}{
\itemize{
\item \code{table}: S3 method for class "table"

\item \code{matrix}: S3 method for class "matrix"

\item \code{numeric}: S3 method for class "numeric"

\item \code{character}: S3 method for class "character"

\item \code{list}: S3 method for class "list"
}}

\examples{
# Gene lists 'atlas' and 'sanger' in 'Cangenes' dataset. Table of joint enrichment
# of GO items in ontology BP at level 3.
tab_atlas.sanger_BP3
dSorensen(tab_atlas.sanger_BP3)
seSorensen(tab_atlas.sanger_BP3)

# Contingency table as a numeric vector:
seSorensen(c(56, 1, 30, 47))
seSorensen(c(56, 1, 30))

?pbtGeneLists
# (Time consuming:)
seSorensen(pbtGeneLists[[2]], pbtGeneLists[[4]],
           listNames = names(pbtGeneLists)[c(2,4)],
           onto = "CC", GOLevel = 5,
           geneUniverse = humanEntrezIDs, orgPackg = "org.Hs.eg.db")
# Essentially, the above code makes the same as:
tab.IRITD3vsKT1 <- buildEnrichTable(pbtGeneLists[[2]], pbtGeneLists[[4]],
                                    listNames = names(pbtGeneLists)[c(2,4)],
                                    onto = "CC", GOLevel = 5,
                                    geneUniverse = humanEntrezIDs, orgPackg = "org.Hs.eg.db")
tab.IRITD3vsKT1
seSorensen(tab.IRITD3vsKT1)

# All pairwise standard errors (quite time consuming):
seSorensen(pbtGeneLists,
           onto = "CC", GOLevel = 5,
           geneUniverse = humanEntrezIDs, orgPackg = "org.Hs.eg.db")

}
\seealso{
\code{\link{nice2x2Table}} for checking and reformatting data,
\code{\link{dSorensen}} for computing the Sorensen-Dice dissimilarity,
\code{\link{duppSorensen}} for the upper limit of a one-sided confidence interval
of the dissimilarity, \code{\link{equivTestSorensen}} for an equivalence test.
}
