% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgrade}
\alias{upgrade}
\alias{upgrade.equivSDhtest}
\alias{upgrade.equivSDhtestList}
\alias{upgrade.AllEquivSDhtest}
\title{Update the result of an Sorensen-Dice equivalence test.}
\usage{
upgrade(x, ...)

\method{upgrade}{equivSDhtest}(x, ...)

\method{upgrade}{equivSDhtestList}(x, ...)

\method{upgrade}{AllEquivSDhtest}(x, ...)
}
\arguments{
\item{x}{an object of class "equivSDhtest", "equivSDhtestList" or "AllEquivSDhtest"}

\item{...}{additional parameters for functions "equivTestSorensen" or "allEquivTestSorensen",
for the moment, these are 'd0', the equivalence limit for the Sorensen-Dice dissimilarity,
and 'conf.level' the confidence level of the one-sided confidence interval upon which the
test is based.}
}
\description{
Change d0 or conf.level in an object of class "equivSDhtest", "equivSDhtestList" or "AllEquivSDhtest"
(i.e.,the output of functions "equivTestSorensen" or "allEquivTestSorensen")
}
\section{Methods (by class)}{
\itemize{
\item \code{equivSDhtest}: S3 method for class "equivSDhtest"

\item \code{equivSDhtestList}: S3 method for class "equivSDhtestList"

\item \code{AllEquivSDhtest}: S3 method for class "allEquivSDhtest"
}}

\examples{
# Result of the equivalence test between gene lists 'waldman' and 'atlas', in dataset
'allOncoGeneLists', at level 4 of the BP ontology:
waldman_atlas.BP.4
class(waldman_atlas.BP.4)
# This may correspond to the result of code like:
# waldman_atlas.BP.4 <- equivTestSorensen(allOncoGeneLists[["waldman"]], allOncoGeneLists[["atlas"]],
#                                         geneUniverse = humanEntrezIDs, orgPackg = "org.Hs.eg.db",
#                                         onto = "BP", GOLevel = 4, listNames = c("waldman", "atlas"))
upgrade(waldman_atlas.BP.4, d0 = 1/(1 + 10/9)) # d0 = 0.4737
upgrade(waldman_atlas.BP.4, d0 = 1/(1 + 2*1.25)) # d0 = 0.2857
upgrade(waldman_atlas.BP.4, d0 = 1/(1 + 2*1.25), conf.level = 0.99)

# All pairwise equivalence tests at level 4 of the BP ontology
BP.4
class(BP.4)
# This may correspond to a call like:
# BP.4 <- equivTestSorensen(allOncoGeneLists,
#                           geneUniverse = humanEntrezIDs, orgPackg = "org.Hs.eg.db",
#                           onto = "BP", GOLevel = 4)
upgrade(BP.4, d0 = 1/(1 + 2*1.25)) # d0 = 0.2857

cancerEquivSorensen
class(cancerEquivSorensen)
upgrade(cancerEquivSorensen, d0 = 1/(1 + 2*1.25)) # d0 = 0.2857

}
